feat: M4 可換式瞄準鏡（五種鏡＋機械瞄具對準）
為什麼：
- 想把 M4 的瞄準鏡拆下來，自己選要裝哪一種。評估時查到兩個原生機制可以直接用：Item.metaInt 原生只有樂透彩券在寫，會存檔也會網路同步；背包拖曳時，被拖物品的 useItemOnItemType 不是 0 就會在房主端呼叫 ItemManager.UseItemOnItem，而原生只處理 1（修理）和 2（重鑄）。

改了什麼：
- M4Scopes：六種鏡的設定（不裝／紅點／全像／ACOG／黃銅鏡／狙擊鏡），metaInt 低位元組編碼（0 = 出廠 ACOG，舊存檔不變），以及安裝／拆卸／拆疊的決策。
- M4Optics（M4Armory partial）：註冊五種鏡物品 47932–47936（各 1 元，useItemOnItemType = 40）；UseItemOnItem prefix 在目標是 M4 時換鏡，原本的鏡放回背包、滿了就掉在腳邊；按 U 由房主驗證後拆鏡（T 是原生定位鍵）；每 3 秒與換鏡時廣播自己的鏡，讓其他玩家的第三人稱模型跟著切換；安裝結果提示。
- M4Model：可切換鏡的 role、後照門立起／折下，以及 ADS 對準點（放在 Recoil 外，後座力仍看得到）。掉落物依 metaInt 顯示鏡。
- MusketScope：M4 依鏡的設定決定倍率段數。1× 瞄具在 LateUpdate 平移旋轉 FPView/Hands，讓照門、紅點、全像窗口對準畫面中心，並畫紅點／全像圈；狙擊鏡加 mil-dot。ScopeCycle 改成 float 段數（1.3×／1.5×）。
- 模型產生器：新增紅點（空心鏡筒）、全像（開放窗口）、黃銅鏡、狙擊鏡與鬼環照門；渲染器新增鏡種對照圖、從眼睛看出去的對準畫面，以及五種鏡的圖示。
- 測試：123 項規則、9,870 項模型（照門中心＝準星尖端＝SightY）、133 項 API 檢查（拖曳路徑、metaInt 寫入者、fpHands 使用者）。
- 版本升為 0.15.0。

+++ GunScope/ScopeMath.cs
@@ -5,15 +5,15 @@ namespace TonyMods
     // Aim opens the first stage, each press advances, and a press on the last stage closes.
     internal sealed class ScopeCycle
     {
-        private readonly int[] stages;
+        private readonly float[] stages;
         private int index = -1;
         public ScopeCycle() : this(3, 6) { }
-        public ScopeCycle(params int[] stages)
+        public ScopeCycle(params float[] stages)
         {
             if (stages == null || stages.Length == 0) throw new ArgumentException("Scope needs at least one stage");
-            this.stages = (int[])stages.Clone();
+            this.stages = (float[])stages.Clone();
         }
-        public int Magnification { get { return index < 0 ? 1 : stages[index]; } }
+        public float Magnification { get { return index < 0 ? 1 : stages[index]; } }
         public bool IsActive { get { return index >= 0; } }
         public bool IsLastStage { get { return index == stages.Length - 1; } }
         public void Advance() { index = index + 1 >= stages.Length ? -1 : index + 1; }
